v2 linker: partial progress

+++ lnasm/src/main/java/com/lnasm/compiler/linker/ILabelResolver.java
@@ -1,7 +1,5 @@
 package com.lnasm.compiler.linker;
 
-import com.lnasm.compiler.linker.ILabelSectionLocator;
-
 public interface ILabelResolver extends ILabelSectionLocator {
-    short resolve(String label);
+    int resolve(String label);
 }
